+++ yagf/src/qsnippet.h
@@ -33,8 +33,8 @@ class QSnippet : public QListWidgetItem
 public:
     explicit QSnippet(QListWidget *parent = 0);
     bool addFile(const QString &name, const QImage * image = NULL);
-    void setRotation(int rot);
-    int getRotation();
+    void setRotation(qreal rot);
+    qreal getRotation();
     void setScale(double sc);
     double getScale();
     void setCrop1(const QRect &rect);
@@ -48,7 +48,7 @@ signals:
 public slots:
 private:
     BolockList blockList;
-    int rotation;
+    qreal rotation;
     double scale;
     QString name;
     QRect crop1;
